fix: move attn impl validation into fsdp utils
